Created SnowplowConstant class or equivalent (closes #35)

+++ Snowplow/SPEmitter.h
@@ -89,7 +89,7 @@ enum SPRequestOptions {
  * Set the buffer time interval to send the events if the buffer hasn't reached it's max capacity yet.
  * @param userTime An int value in seconds
  */
-- (void) setNewBufferTime:(int) userTime;
+- (void) setNewBufferTime:(NSInteger) userTime;
 
 /**
  * Returns the total Database Count
